refactor: CrewService에서 DemandService 분리
