+++ src/test/java/com/retrip/crew/application/in/CrewServiceTest.java
@@ -1,57 +1,34 @@
 package com.retrip.crew.application.in;
 
-import static com.retrip.crew.common.fixture.CrewFixture.LEADER_ID;
-import static com.retrip.crew.common.fixture.CrewFixture.MEMBER_ID;
-import static com.retrip.crew.common.fixture.CrewFixture.createCrew;
-import static com.retrip.crew.common.fixture.CrewFixture.createCrewRequest;
-import static com.retrip.crew.common.fixture.CrewFixture.createCrewWithMembers;
-import static com.retrip.crew.common.fixture.CrewFixture.createMultipleCrews;
-import static com.retrip.crew.common.fixture.CrewFixture.정수_ID;
-import static com.retrip.crew.common.fixture.CrewFixture.준호_ID;
-import static com.retrip.crew.common.fixture.CrewFixture.지수_ID;
-import static com.retrip.crew.common.fixture.CrewFixture.혁진_ID;
-import static com.retrip.crew.common.fixture.CrewFixture.홍석_ID;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
 import com.retrip.crew.application.in.request.IntroductionCreateRequest;
 import com.retrip.crew.application.in.request.IntroductionDeleteRequest;
 import com.retrip.crew.application.in.request.IntroductionUpdateRequest;
 import com.retrip.crew.application.in.request.crew.CrewCreateRequest;
 import com.retrip.crew.application.in.request.crew.CrewOrder;
 import com.retrip.crew.application.in.request.crew.CrewUpdateRequest;
-import com.retrip.crew.application.in.request.demand.CreateDemandRequest;
-import com.retrip.crew.application.in.request.demand.DemandOrder;
 import com.retrip.crew.application.in.response.IntroductionCreateResponse;
 import com.retrip.crew.application.in.response.IntroductionDetailResponse;
 import com.retrip.crew.application.in.response.crew.CrewCreateResponse;
 import com.retrip.crew.application.in.response.crew.CrewDetailResponse;
 import com.retrip.crew.application.in.response.crew.CrewListResponse;
 import com.retrip.crew.application.in.response.crew.CrewUpdateResponse;
-import com.retrip.crew.application.in.response.demand.CreateDemandResponse;
-import com.retrip.crew.application.in.response.demand.CrewsOfDemandResponse;
-import com.retrip.crew.application.in.response.demand.DemandsResponse;
 import com.retrip.crew.common.ServiceTest;
 import com.retrip.crew.domain.entity.Crew;
 import com.retrip.crew.domain.entity.CrewMemberRole;
-import com.retrip.crew.domain.entity.Demand;
 import com.retrip.crew.domain.entity.Introduction;
-import com.retrip.crew.domain.exception.DuplicateDemandException;
 import com.retrip.crew.domain.exception.common.InvalidAccessException;
-import com.retrip.crew.domain.vo.DemandStatus;
 import com.retrip.crew.infra.adapter.in.presentation.rest.common.ScrollPageResponse;
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.test.util.ReflectionTestUtils;
+
+import java.util.List;
+import java.util.UUID;
+
+import static com.retrip.crew.common.fixture.CrewFixture.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class CrewServiceTest extends ServiceTest {
     @Test
@@ -90,101 +67,6 @@ class CrewServiceTest extends ServiceTest {
                 () -> assertThat(response.title()).isEqualTo("강릉 크루원 구함"),
                 () -> assertThat(response.maxMembers()).isEqualTo(200)
         );
-    }
-
-    @Test
-    void 크루_참여_요청을_생성한다() {
-        // given
-        Crew crew = crewRepository.save(createCrew(LEADER_ID));
-        CreateDemandRequest request = new CreateDemandRequest(MEMBER_ID);
-
-        // when
-        CreateDemandResponse response = crewService.createDemand(crew.getId(), request);
-
-        // then
-        List<Demand> demands = crew.getRecruitment().getDemands();
-        assertAll(
-                () -> assertThat(demands.size()).isEqualTo(1),
-                () -> assertThat(response.memberId()).isEqualTo(demands.get(0).getMemberId())
-        );
-    }
-
-    @Test
-    void 이미_요청한_사용자는_다시_크루에_요청할_수_없다() {
-        // given
-        Crew crew = createCrew(LEADER_ID);
-        crew.demand(MEMBER_ID);
-        crew.demand(UUID.randomUUID());
-        crew.demand(UUID.randomUUID());
-        Crew save = crewRepository.save(crew);
-        CreateDemandRequest request = new CreateDemandRequest(MEMBER_ID);
-
-        // when, then
-        assertThatThrownBy(() -> crewService.createDemand(save.getId(), request))
-                .isExactlyInstanceOf(DuplicateDemandException.class);
-    }
-
-    @Test
-    void 크루_참여_요청을_취소한다() {
-        // given
-        Crew crew = crewRepository.save(createCrew(LEADER_ID));
-        Demand demand = crew.demand(MEMBER_ID);
-
-        // when
-        crewService.cancelDemand(crew.getId(), demand.getId(), MEMBER_ID);
-
-        // then
-        assertThat(demand.getStatus()).isEqualTo(DemandStatus.CANCELED);
-    }
-
-    @ParameterizedTest
-    @CsvSource({"PENDING,2", "APPROVED,2", "REJECTED,1"})
-    void 리더가_크루_요청_목록을_조회한다(String status, int expected) {
-        // given
-        Crew crew = crewRepository.save(createCrew(LEADER_ID));
-        Demand 정수 = new Demand(정수_ID, crew);
-        Demand 홍석 = new Demand(홍석_ID, crew);
-        Demand 준호 = new Demand(준호_ID, crew);
-        Demand 지수 = new Demand(지수_ID, crew);
-        Demand 혁진 = new Demand(혁진_ID, crew);
-        List<Demand> demands = List.of(정수, 홍석, 준호, 지수, 혁진);
-
-        ReflectionTestUtils.setField(준호, "status", DemandStatus.APPROVED);
-        ReflectionTestUtils.setField(지수, "status", DemandStatus.APPROVED);
-        ReflectionTestUtils.setField(혁진, "status", DemandStatus.REJECTED);
-        ReflectionTestUtils.setField(crew.getRecruitment(), "demands", demands);
-
-        Pageable pageable = PageRequest.of(0, 10);
-
-        // when
-        Page<DemandsResponse> response =
-                crewService.getDemands(crew.getId(), LEADER_ID, status, pageable, DemandOrder.DATE, "desc");
-
-        // then
-        assertThat(response.getTotalElements()).isEqualTo(expected);
-    }
-
-    @Test
-    void 참여_요청자가_속한_크루_목록을_조회한다() {
-        // given
-        Crew crew = createCrew(LEADER_ID);
-        Demand demand = new Demand(홍석_ID, crew);
-        ReflectionTestUtils.setField(crew.getRecruitment(), "demands", List.of(demand));
-
-        crewRepository.save(crew);
-        crewRepository.save(createCrewWithMembers(MEMBER_ID));
-        crewRepository.save(createCrewWithMembers(MEMBER_ID));
-        crewRepository.save(createCrewWithMembers(MEMBER_ID));
-        crewRepository.save(createCrewWithMembers(MEMBER_ID));
-        crewRepository.save(createCrewWithMembers(MEMBER_ID));
-        Pageable pageable = PageRequest.of(0, 10);
-
-        // when
-        Page<CrewsOfDemandResponse> response =
-                crewService.getCrewsOfDemand(crew.getId(), demand.getId(), LEADER_ID, pageable, CrewOrder.DATE, "desc");
-
-        // then
-        assertThat(response.getTotalElements()).isEqualTo(5);
     }
 
     @Test
