+++ src/test/java/com/retrip/crew/common/ServiceTest.java
@@ -1,14 +1,9 @@
 package com.retrip.crew.common;
 
 import com.retrip.crew.application.in.CrewService;
-import com.retrip.crew.application.out.repository.CrewDemandRepository;
-import com.retrip.crew.application.out.repository.CrewMemberRepository;
-import com.retrip.crew.application.out.repository.CrewQueryRepository;
-import com.retrip.crew.application.out.repository.CrewRepository;
-import com.retrip.crew.application.out.repository.IntroductionQueryRepository;
-import com.retrip.crew.application.out.repository.IntroductionRepository;
+import com.retrip.crew.application.in.DemandService;
+import com.retrip.crew.application.out.repository.*;
 import com.retrip.crew.common.config.QuerydslConfig;
-import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
@@ -44,8 +39,11 @@ public class ServiceTest {
 
     protected CrewService crewService;
 
+    protected DemandService demandService;
+
     @BeforeEach
     void setUp() {
-        crewService = new CrewService(crewRepository, crewMemberRepository, crewQueryRepository, demandRepository,  introductionRepository, introductionQueryRepository);
+        crewService = new CrewService(crewRepository, crewMemberRepository, crewQueryRepository,  introductionRepository, introductionQueryRepository);
+        demandService = new DemandService(crewRepository, demandRepository, crewQueryRepository);
     }
 }
